refactor: group LoginController depencies in a single file
