+++ src/presentation/controllers/login/login.ts
@@ -1,8 +1,6 @@
-import { Authentication } from '../../../domain/usecases/authentication'
+import { EmailValidator, Authentication, Controller, HttpRequest, HttpResponse } from './login-protocols'
 import { InvalidParamError, MissingParamError } from '../../erros'
 import { badRequest, serverError, unauthorized } from '../../helpers/http-helper'
-import { Controller, HttpRequest, HttpResponse } from '../../protocols'
-import { EmailValidator } from '../signup/signup-protocols'
 
 export class LoginController implements Controller {
   constructor (private readonly emailValidator: EmailValidator, private readonly authentication: Authentication) {}
